+++ src/graph_pes/__init__.py
@@ -1,5 +1,10 @@
+import warnings
+
 from graph_pes.atomic_graph import AtomicGraph
 from graph_pes.graph_pes_model import GraphPESModel
+
+# hide the annoying FutureWarning from e3nn
+warnings.filterwarnings("ignore", category=FutureWarning, module="e3nn")
 
 __all__ = ["AtomicGraph", "GraphPESModel"]
 __version__ = "0.0.31"